requirements.txt
# Flask web framework
flask==3.0.0
gunicorn==21.2.0

# HTTP requests (for Mealie API)
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# JSON schema validation
jsonschema==4.20.0

# Development tools
ruff==0.3.4
pytest==8.1.1
pytest-cov==4.1.0
